# api/_gen/swagger/swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
...
		"GamePlayer": {
			"properties": {
				"steamId": {
					"type": "string"
				},
				"civType": {
					"type": "string"
				},
				"hasSurrendered": {
					"type": "boolean"
				},
				"turnsPlayed": {
					"type": "number",
					"format": "double"
				},
				"turnsSkipped": {
					"type": "number",
					"format": "double"
				},
				"timeTaken": {
					"type": "number",
					"format": "double"
				},
				"fastTurns": {
					"type": "number",
					"format": "double"
				},
				"slowTurns": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"steamId",
				"civType"
			],
			"type": "object"
		},
		"Game": {
			"properties": {
				"gameId": {
					"type": "string"
				},
				"createdBySteamId": {
					"type": "string"
				},
				"inProgress": {
					"type": "boolean"
				},
				"hashedPassword": {
					"type": "string"
				},
				"players": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/GamePlayer"
					}
				},
				"discourseTopicId": {
					"type": "number",
					"format": "double"
				},
				"currentPlayerSteamId": {
					"type": "string"
				},
				"turnTimerMinutes": {
					"type": "number",
					"format": "double"
				},
				"round": {
					"type": "number",
					"format": "double"
				},
				"gameTurnRangeKey": {
					"type": "number",
					"format": "double"
				},
				"completed": {
					"type": "boolean"
				},
				"createdAt": {
					"type": "string",
					"format": "date-time"
				},
				"updatedAt": {
					"type": "string",
					"format": "date-time"
				},
				"version": {
					"type": "number",
					"format": "double"
				},
				"displayName": {
					"type": "string"
				},
				"description": {
					"type": "string"
				},
				"dlc": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"slots": {
					"type": "number",
					"format": "double"
				},
				"humans": {
					"type": "number",
					"format": "double"
				},
				"gameSpeed": {
					"type": "string"
				},
				"mapFile": {
					"type": "string"
				},
				"mapSize": {
					"type": "string"
				},
				"allowJoinAfterStart": {
					"type": "boolean"
				}
			},
			"required": [
				"gameId",
				"createdBySteamId",
				"players",
				"currentPlayerSteamId",
				"displayName",
				"description",
				"dlc",
				"slots",
				"humans",
				"gameSpeed",
				"mapFile",
				"mapSize"
			],
			"type": "object"
		},
		"ErrorResponse": {
			"properties": {
				"message": {
					"type": "string"
				}
			},
			"required": [
				"message"
			],
			"type": "object"
		},
		"ChangeCivRequestBody": {
			"properties": {
				"playerCiv": {
					"type": "string"
				}
			},
			"required": [
				"playerCiv"
			],
			"type": "object"
		},
		"CreateGameRequestBody": {
			"properties": {
				"player1Civ": {
					"type": "string"
				},
				"password": {
					"type": "string"
				},
				"displayName": {
					"type": "string"
				},
				"description": {
					"type": "string"
				},
				"dlc": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"slots": {
					"type": "number",
					"format": "double"
				},
				"humans": {
					"type": "number",
					"format": "double"
				},
				"gameSpeed": {
					"type": "string"
				},
				"mapFile": {
					"type": "string"
				},
				"mapSize": {
					"type": "string"
				},
				"allowJoinAfterStart": {
					"type": "boolean"
				}
			},
			"required": [
				"player1Civ",
				"password",
				"displayName",
				"description",
				"dlc",
				"slots",
				"humans",
				"gameSpeed",
				"mapFile",
				"mapSize"
			],
			"type": "object"
		},
		"GameRequestBody": {
			"properties": {
				"password": {
					"type": "string"
				},
				"displayName": {
					"type": "string"
				},
				"description": {
					"type": "string"
				},
				"dlc": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"slots": {
					"type": "number",
					"format": "double"
				},
				"humans": {
					"type": "number",
					"format": "double"
				},
				"gameSpeed": {
					"type": "string"
				},
				"mapFile": {
					"type": "string"
				},
				"mapSize": {
					"type": "string"
				},
				"allowJoinAfterStart": {
					"type": "boolean"
				}
			},
			"required": [
				"password",
				"displayName",
				"description",
				"dlc",
				"slots",
				"humans",
				"gameSpeed",
				"mapFile",
				"mapSize"
			],
			"type": "object"
		},
		"JoinGameRequestBody": {
			"properties": {
				"password": {
					"type": "string"
				},
				"playerCiv": {
					"type": "string"
				}
			},
			"required": [
				"password",
				"playerCiv"
			],
			"type": "object"
		},
		"OpenGamesResponse": {
			"properties": {
				"notStarted": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Game"
					}
				},
				"openSlots": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Game"
					}
				}
			},
			"required": [
				"notStarted",
				"openSlots"
			],
			"type": "object"
		},
		"SurrenderBody": {
			"properties": {
				"kickUserId": {
					"type": "string"
				}
			},
			"required": [
				"kickUserId"
			],
			"type": "object"
		},
...
	},
	"info": {
		"title": "serverless-api",
		"version": "1.0.0"
	},
	"paths": {
...
		"/game/{gameId}/changeCiv": {
			"post": {
				"operationId": "GameChangeCiv",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Game"
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ChangeCivRequestBody"
						}
					}
				]
			}
		},
		"/game/create": {
			"post": {
				"operationId": "GameCreate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Game"
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/CreateGameRequestBody"
						}
					}
				]
			}
		},
		"/game/{gameId}/delete": {
			"post": {
				"operationId": "GameDelete",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/game/{gameId}/edit": {
			"post": {
				"operationId": "GameEdit",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Game"
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/GameRequestBody"
						}
					}
				]
			}
		},
		"/game/{gameId}/join": {
			"post": {
				"operationId": "GameJoin",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Game"
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/JoinGameRequestBody"
						}
					}
				]
			}
		},
		"/game/{gameId}/leave": {
			"post": {
				"operationId": "GameLeave",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Game"
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/game/listOpen": {
			"get": {
				"operationId": "GameListOpen",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/OpenGamesResponse"
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/game/{gameId}/start": {
			"post": {
				"operationId": "GameStart",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Game"
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/game/{gameId}/surrender": {
			"post": {
				"operationId": "GameSurrender",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Game"
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					}
				},
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/SurrenderBody"
						}
					}
				]
			}
		},
		"/game/{gameId}": {
			"get": {
				"operationId": "GameGet",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Game"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"type": "string"
					}
				]
			}
		},
...
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		},
		"basic_auth_internal": {
			"type": "basic"
		}
	},
	"host": "localhost:3000"
}